// loadtest/api-client/generated/models/DBiamPersonResponse.ts
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DBiamPersonenkontextResponse } from '../models/DBiamPersonenkontextResponse.ts';
import { PersonResponse } from '../models/PersonResponse.ts';
import { HttpFile } from '../http/http.ts';

export class DBiamPersonResponse {
    'person': PersonResponse;
    'dBiamPersonenkontextResponses': Array<DBiamPersonenkontextResponse>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "person",
            "baseName": "person",
            "type": "PersonResponse",
            "format": ""
        },
        {
            "name": "dBiamPersonenkontextResponses",
            "baseName": "dBiamPersonenkontextResponses",
            "type": "Array<DBiamPersonenkontextResponse>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DBiamPersonResponse.attributeTypeMap;
    }

    public constructor() {
    }
}