// loadtest/api-client/generated/models/OrganisationsTyp.ts
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http.ts';

export enum OrganisationsTyp {
    Root = 'ROOT',
    Land = 'LAND',
    Traeger = 'TRAEGER',
    Schule = 'SCHULE',
    Klasse = 'KLASSE',
    Anbieter = 'ANBIETER',
    SonstigeOrganisationEinrichtung = 'SONSTIGE ORGANISATION / EINRICHTUNG',
    Unbestaetigt = 'UNBESTAETIGT'
}