// loadtest/api-client/generated/models/RollenArt.ts
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from "../http/http.ts";

export enum RollenArt {
  Lern = "LERN",
  Lehr = "LEHR",
  Extern = "EXTERN",
  Orgadmin = "ORGADMIN",
  Leit = "LEIT",
  Sysadmin = "SYSADMIN",
}